requirements.txt
Quart==0.10.0
kafka-python==1.4.7
python-dotenv==0.10.3
fake-rpi==0.6.2